requirements.txt
aiohttp
Pillow
playwright
pyquery
requests
tqdm
yapf
fire
lzstring
loguru
yapf